Preserve all namespace prefixes
The motivation for this is to keep output XML as close to input as
possible. This improves compatibility if a client relies on specific
namespace prefixes like 'itunes'.

To achieve this, a modified TreeBuilder is used as a hook to record
namespaces in the XML stream. This is then used when writing out the XML
to serve.

There are several downsides to this:

1. The only way to make ElementTree set namespaces prefixes in output is
   a global registry. This means there is a race if multiple threads are
   writing XML output at the same time. To resolve this, we change our
   Dockerfile to only run with one thread.

2. ElementTree provides no way to remove entries from the global
   namespace registry. Because of this we have ugly monkeypatching going
   on.