Tighter QueryBuilder validation
The jQuery QueryBuilder UI used in this project is a lot stricter
than the python implementation. It is desirable to prevent python
generating QueryBuilder objects (e.g. in the CLI) that will cause
issues in the UI.

This therefore tightens up the validation in python to (closer)
parity with the UI, and tests it.

This neccessitates moving the dataclass to a separate file to avoid
a circular import.